Conflicts:
	src/org.eclipse.ice.viz.service/src/org/eclipse/ice/viz/service/BasicVizServiceFactory.java
	src/org.eclipse.ice.viz.service/src/org/eclipse/ice/viz/service/csv/CSVPlot.java

+++ src/org.eclipse.ice.client.widgets.reactoreditor.grid/src/org/eclipse/ice/client/widgets/reactoreditor/grid/GridEditorTools.java
@@ -313,7 +313,7 @@ public class GridEditorTools {
 
 					// Use the law of cosines to determine the distance of the
 					// hexagon's center from the origin.
-					double R2 = (double) i2 + (double) r2 - 2 * i * r
+					double R2 = i2 + r2 - 2 * i * r
 							* cosPiDivThree;
 					double R = Math.sqrt(R2);
 
@@ -379,7 +379,7 @@ public class GridEditorTools {
 
 					// Use the law of cosines to determine the distance of the
 					// hexagon's center from the origin.
-					double R2 = (double) i2 + (double) r2 - 2 * i * r
+					double R2 = i2 + r2 - 2 * i * r
 							* cosPiDivThree;
 					double R = Math.sqrt(R2);
